+++ NihavGameEngine/source/Library.Shared/HeapManager.h
@@ -20,11 +20,11 @@ namespace Library
 		static void DestroyHeapManager();
 
 		void Init(const std::vector<HeapDef>& heapDefs, std::uint32_t defaultHeapId);
-		std::uint32_t GetHeapSize(std::uint32_t id) const;
+		std::size_t GetHeapSize(std::uint32_t id) const;
 		const std::string& GetHeapName(std::uint32_t id) const;
 		std::uint32_t GetHeapNumBlocks(std::uint32_t id) const;
 
-		void* AllocateBlock(std::uint32_t heapId, std::uint32_t bytes, const std::string& filename, std::uint32_t linenumber, const std::string& tag, const std::chrono::high_resolution_clock::time_point& time, std::uint32_t alignment = 4);
+		void* AllocateBlock(std::uint32_t heapId, std::size_t bytes, const std::string& filename, std::uint32_t linenumber, const std::string& tag, const std::chrono::high_resolution_clock::time_point& time, std::uint32_t alignment = 4);
 		void FreeBlock(void* ptr);
 
 		void SetCurrentHeapID(std::uint32_t id);
